task updated

+++ 0x07-pointers_arrays_strings/1-memcpy.c
@@ -2,19 +2,20 @@
 
 /**
  * _memcpy - copies memory area
- * @dest: destination of the copied memory area
- * @src: source of the copied memory area
- * @n: number bytes to be copied from src
- * Return: return a pointer to a dest
+ * @dest: destination of the copied memory
+ * @src: source from which memory is copied
+ * @n: number of bytes to be copied from src
+ * Return: returns a char pointer dest
  */
 
 char *_memcpy(char *dest, char *src, unsigned int n)
 {
-	int i;
+	unsigned int i;
 
-	for (i = 0; i < (int)n; i++)
+	for (i = 0; i < n; i++)
 	{
 		*(dest + i) = *(src + i);
 	}
 	return (dest);
 }
+
